requirements.txt
async-timeout==5.0.1
blinker==1.9.0
click==8.3.0
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
redis==7.0.1
Werkzeug==3.1.3
openai>=1.0.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
tenacity>=8.2.0

